Fix editor on default gitconfig to be vim to avoid problems
The fallback to `$EDITOR` is not working now with the `-w` flag as it
used to.
Fix those so that I can use `e .` and similar commands without getting a
`line 16: exec: atom -w: not found`.

We can always change the default editor for git commits to be atom by
doing: `git config --global core.editor "atom -w -n"`